requirements.txt
aiogram==3.0.0b7
python-dotenv
requests